Address review: manual sitemap, extract JSON-LD component, drop comments
- sitemap.ts: revert to a simple manual route list (per review — the
  filesystem auto-discovery was unnecessarily complex). Keeps all current
  top-level routes.
- page.tsx: move the Person/WebSite JSON-LD into a SiteStructuredData
  component.
- seo.tsx: remove two redundant comments.
